Created 'pqappstoreservices' and moved classes from 'pqlib.AppStoreServices' into 'pqappstoreservices' ('pqlib' has been removed from the project)
