+++ pqappstoreservices/Interfaces/IAppUpdater.cs
@@ -1,11 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
 using Windows.Services.Store;
 
-namespace pqlib.AppStoreServices
+namespace pqappstoreservices.Interfaces
 {
     public interface IAppUpdater
     {
