+++ pqappstoreservices/pqappstoreservices.csproj
@@ -4,11 +4,11 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{489D68FC-AF00-4E3D-B05A-05CFF491F73F}</ProjectGuid>
+    <ProjectGuid>{8AE13322-9FFA-44CA-9B3E-08B01CC03BA9}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>pqytparser</RootNamespace>
-    <AssemblyName>pqytparser</AssemblyName>
+    <RootNamespace>pqappstoreservices</RootNamespace>
+    <AssemblyName>pqappstoreservices</AssemblyName>
     <DefaultLanguage>en-US</DefaultLanguage>
     <TargetPlatformIdentifier>UAP</TargetPlatformIdentifier>
     <TargetPlatformVersion Condition=" '$(TargetPlatformVersion)' == '' ">10.0.16299.0</TargetPlatformVersion>
@@ -106,33 +106,15 @@
     <RestoreProjectStyle>PackageReference</RestoreProjectStyle>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="Interfaces\IVideoInfoDomDecoder.cs" />
-    <Compile Include="Interfaces\IVideoInfoParser.cs" />
-    <Compile Include="Models\VideoDownloadInfo.cs" />
-    <Compile Include="Models\VideoMetadata.cs" />
-    <Compile Include="ViewModels\VideoInfoDomDecoder.cs" />
-    <Compile Include="ViewModels\VideoInfoParser.cs" />
+    <Compile Include="Interfaces\IAppUpdater.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Resources\FileExtensions.cs" />
-    <Compile Include="Resources\MimeTypeEnum.cs" />
-    <Compile Include="Resources\MediaQualityEnum.cs" />
-    <Compile Include="Resources\FileTypeEnum.cs" />
-    <Compile Include="Resources\MimeTypes.cs" />
-    <Compile Include="Resources\SuggestedPlaybackQualityEnum.cs" />
-    <Compile Include="Resources\VideoAvailabilityEnum.cs" />
-    <EmbeddedResource Include="Properties\pqytparser.rd.xml" />
+    <Compile Include="ViewModels\AppUpdater.cs" />
+    <EmbeddedResource Include="Properties\pqappstoreservices.rd.xml" />
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.NETCore.UniversalWindowsPlatform">
-      <Version>6.0.7</Version>
+      <Version>6.0.8</Version>
     </PackageReference>
-  </ItemGroup>
-  <ItemGroup />
-  <ItemGroup>
-    <ProjectReference Include="..\pqreflection\pqreflection.csproj">
-      <Project>{7eb211b5-dc0a-48c5-9657-6e1ce654dfcf}</Project>
-      <Name>pqreflection</Name>
-    </ProjectReference>
   </ItemGroup>
   <PropertyGroup Condition=" '$(VisualStudioVersion)' == '' or '$(VisualStudioVersion)' &lt; '14.0' ">
     <VisualStudioVersion>14.0</VisualStudioVersion>
