+++ projectquail.sln
@@ -7,13 +7,13 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "projectquail", "projectquai
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "pqytparser", "pqytparser\pqytparser.csproj", "{489D68FC-AF00-4E3D-B05A-05CFF491F73F}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "pqlib", "pqlib\pqlib.csproj", "{D064263D-1601-4B2D-959F-B208DC2167EA}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "projectquailunittest", "projectquailunittest\projectquailunittest.csproj", "{56188950-DF00-4F3B-A598-56FD6B9A15DD}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "pqcommonregex", "pqcommonregex\pqcommonregex.csproj", "{40B78344-38C5-44C8-B2C1-40C0A5B37629}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "pqreflection", "pqreflection\pqreflection.csproj", "{7EB211B5-DC0A-48C5-9657-6E1CE654DFCF}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "pqappstoreservices", "pqappstoreservices\pqappstoreservices.csproj", "{8AE13322-9FFA-44CA-9B3E-08B01CC03BA9}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -63,22 +63,6 @@ Global
 		{489D68FC-AF00-4E3D-B05A-05CFF491F73F}.Release|x64.Build.0 = Release|x64
 		{489D68FC-AF00-4E3D-B05A-05CFF491F73F}.Release|x86.ActiveCfg = Release|x86
 		{489D68FC-AF00-4E3D-B05A-05CFF491F73F}.Release|x86.Build.0 = Release|x86
-		{D064263D-1601-4B2D-959F-B208DC2167EA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D064263D-1601-4B2D-959F-B208DC2167EA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D064263D-1601-4B2D-959F-B208DC2167EA}.Debug|ARM.ActiveCfg = Debug|ARM
-		{D064263D-1601-4B2D-959F-B208DC2167EA}.Debug|ARM.Build.0 = Debug|ARM
-		{D064263D-1601-4B2D-959F-B208DC2167EA}.Debug|x64.ActiveCfg = Debug|x64
-		{D064263D-1601-4B2D-959F-B208DC2167EA}.Debug|x64.Build.0 = Debug|x64
-		{D064263D-1601-4B2D-959F-B208DC2167EA}.Debug|x86.ActiveCfg = Debug|x86
-		{D064263D-1601-4B2D-959F-B208DC2167EA}.Debug|x86.Build.0 = Debug|x86
-		{D064263D-1601-4B2D-959F-B208DC2167EA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D064263D-1601-4B2D-959F-B208DC2167EA}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D064263D-1601-4B2D-959F-B208DC2167EA}.Release|ARM.ActiveCfg = Release|ARM
-		{D064263D-1601-4B2D-959F-B208DC2167EA}.Release|ARM.Build.0 = Release|ARM
-		{D064263D-1601-4B2D-959F-B208DC2167EA}.Release|x64.ActiveCfg = Release|x64
-		{D064263D-1601-4B2D-959F-B208DC2167EA}.Release|x64.Build.0 = Release|x64
-		{D064263D-1601-4B2D-959F-B208DC2167EA}.Release|x86.ActiveCfg = Release|x86
-		{D064263D-1601-4B2D-959F-B208DC2167EA}.Release|x86.Build.0 = Release|x86
 		{56188950-DF00-4F3B-A598-56FD6B9A15DD}.Debug|Any CPU.ActiveCfg = Debug|x86
 		{56188950-DF00-4F3B-A598-56FD6B9A15DD}.Debug|ARM.ActiveCfg = Debug|ARM
 		{56188950-DF00-4F3B-A598-56FD6B9A15DD}.Debug|ARM.Build.0 = Debug|ARM
@@ -131,6 +115,22 @@ Global
 		{7EB211B5-DC0A-48C5-9657-6E1CE654DFCF}.Release|x64.Build.0 = Release|x64
 		{7EB211B5-DC0A-48C5-9657-6E1CE654DFCF}.Release|x86.ActiveCfg = Release|x86
 		{7EB211B5-DC0A-48C5-9657-6E1CE654DFCF}.Release|x86.Build.0 = Release|x86
+		{8AE13322-9FFA-44CA-9B3E-08B01CC03BA9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8AE13322-9FFA-44CA-9B3E-08B01CC03BA9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8AE13322-9FFA-44CA-9B3E-08B01CC03BA9}.Debug|ARM.ActiveCfg = Debug|ARM
+		{8AE13322-9FFA-44CA-9B3E-08B01CC03BA9}.Debug|ARM.Build.0 = Debug|ARM
+		{8AE13322-9FFA-44CA-9B3E-08B01CC03BA9}.Debug|x64.ActiveCfg = Debug|x64
+		{8AE13322-9FFA-44CA-9B3E-08B01CC03BA9}.Debug|x64.Build.0 = Debug|x64
+		{8AE13322-9FFA-44CA-9B3E-08B01CC03BA9}.Debug|x86.ActiveCfg = Debug|x86
+		{8AE13322-9FFA-44CA-9B3E-08B01CC03BA9}.Debug|x86.Build.0 = Debug|x86
+		{8AE13322-9FFA-44CA-9B3E-08B01CC03BA9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8AE13322-9FFA-44CA-9B3E-08B01CC03BA9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8AE13322-9FFA-44CA-9B3E-08B01CC03BA9}.Release|ARM.ActiveCfg = Release|ARM
+		{8AE13322-9FFA-44CA-9B3E-08B01CC03BA9}.Release|ARM.Build.0 = Release|ARM
+		{8AE13322-9FFA-44CA-9B3E-08B01CC03BA9}.Release|x64.ActiveCfg = Release|x64
+		{8AE13322-9FFA-44CA-9B3E-08B01CC03BA9}.Release|x64.Build.0 = Release|x64
+		{8AE13322-9FFA-44CA-9B3E-08B01CC03BA9}.Release|x86.ActiveCfg = Release|x86
+		{8AE13322-9FFA-44CA-9B3E-08B01CC03BA9}.Release|x86.Build.0 = Release|x86
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
